requirements.txt
streamlit
pillow
torch
transformers
nltk
matplotlib
opencv-python-headless
pandas
requests